Adds back a refactored COD file reader.
This is now in a separate directory, to show that they are not
required to use the simulator.

src/util/ has a dependency on src/, but not vice versa. (I'm not happy
about the naming, but good enough for now.)

This code does not rely on files; it takes a borrowed std::istream.